Crate rnix

Re-exports

pub use self::tokenizer::tokenize;

pub use ast::Root;

Modules

ast

Provides a type system for the AST, in some sense

parser

The parser: turns a series of tokens into an AST

tokenizer

The tokenizer: turns a string into tokens, such as numbers, strings, and keywords

Macros

T

match_ast

Matches a SyntaxNode against an ast type.

Structs

Parse

The result of a parse

TextRange

A range in text, represented as a pair of TextSize.

TextSize

A measure of text length. Also, equivalently, an index into text.

Enums

NixLanguage

NodeOrToken

SyntaxKind

TokenAtOffset

There might be zero, one or two leaves at a given offset.

WalkEvent

WalkEvent describes tree walking process.

Type Aliases

SyntaxElement

SyntaxElementChildren

SyntaxNode

SyntaxNodeChildren

SyntaxToken