rnix

Enum SyntaxKind

#[repr(u16)]
pub enum SyntaxKind {

    TOKEN_COMMENT,
    TOKEN_ERROR,
    TOKEN_WHITESPACE,
    TOKEN_ASSERT,
    TOKEN_ELSE,
    TOKEN_IF,
    TOKEN_IN,
    TOKEN_INHERIT,
    TOKEN_LET,
    TOKEN_OR,
    TOKEN_REC,
    TOKEN_THEN,
    TOKEN_WITH,
    TOKEN_L_BRACE,
    TOKEN_R_BRACE,
    TOKEN_L_BRACK,
    TOKEN_R_BRACK,
    TOKEN_ASSIGN,
    TOKEN_AT,
    TOKEN_COLON,
    TOKEN_COMMA,
    TOKEN_DOT,
    TOKEN_ELLIPSIS,
    TOKEN_QUESTION,
    TOKEN_SEMICOLON,
    TOKEN_L_PAREN,
    TOKEN_R_PAREN,
    TOKEN_CONCAT,
    TOKEN_INVERT,
    TOKEN_UPDATE,
    TOKEN_ADD,
    TOKEN_SUB,
    TOKEN_MUL,
    TOKEN_DIV,
    TOKEN_AND_AND,
    TOKEN_EQUAL,
    TOKEN_IMPLICATION,
    TOKEN_LESS,
    TOKEN_LESS_OR_EQ,
    TOKEN_MORE,
    TOKEN_MORE_OR_EQ,
    TOKEN_NOT_EQUAL,
    TOKEN_OR_OR,
    TOKEN_PIPE_RIGHT,
    TOKEN_PIPE_LEFT,
    TOKEN_FLOAT,
    TOKEN_IDENT,
    TOKEN_INTEGER,
    TOKEN_INTERPOL_END,
    TOKEN_INTERPOL_START,
    TOKEN_PATH,
    TOKEN_URI,
    TOKEN_STRING_CONTENT,
    TOKEN_STRING_END,
    TOKEN_STRING_START,
    NODE_APPLY,
    NODE_ASSERT,
    NODE_ATTRPATH,
    NODE_DYNAMIC,
    NODE_ERROR,
    NODE_IDENT,
    NODE_IF_ELSE,
    NODE_SELECT,
    NODE_INHERIT,
    NODE_INHERIT_FROM,
    NODE_STRING,
    NODE_INTERPOL,
    NODE_LAMBDA,
    NODE_IDENT_PARAM,
    NODE_LEGACY_LET,
    NODE_LET_IN,
    NODE_LIST,
    NODE_BIN_OP,
    NODE_PAREN,
    NODE_PATTERN,
    NODE_PAT_BIND,
    NODE_PAT_ENTRY,
    NODE_ROOT,
    NODE_ATTR_SET,
    NODE_ATTRPATH_VALUE,
    NODE_UNARY_OP,
    NODE_LITERAL,
    NODE_WITH,
    NODE_PATH,
    NODE_HAS_ATTR,
    // some variants omitted

}

Variants

TOKEN_COMMENT

TOKEN_ERROR

TOKEN_WHITESPACE

TOKEN_ASSERT

TOKEN_ELSE

TOKEN_IF

TOKEN_IN

TOKEN_INHERIT

TOKEN_LET

TOKEN_OR

TOKEN_REC

TOKEN_THEN

TOKEN_WITH

TOKEN_L_BRACE

TOKEN_R_BRACE

TOKEN_L_BRACK

TOKEN_R_BRACK

TOKEN_ASSIGN

TOKEN_AT

TOKEN_COLON

TOKEN_COMMA

TOKEN_DOT

TOKEN_ELLIPSIS

TOKEN_QUESTION

TOKEN_SEMICOLON

TOKEN_L_PAREN

TOKEN_R_PAREN

TOKEN_CONCAT

TOKEN_INVERT

TOKEN_UPDATE

TOKEN_ADD

TOKEN_SUB

TOKEN_MUL

TOKEN_DIV

TOKEN_AND_AND

TOKEN_EQUAL

TOKEN_IMPLICATION

TOKEN_LESS

TOKEN_LESS_OR_EQ

TOKEN_MORE

TOKEN_MORE_OR_EQ

TOKEN_NOT_EQUAL

TOKEN_OR_OR

TOKEN_PIPE_RIGHT

TOKEN_PIPE_LEFT

TOKEN_FLOAT

TOKEN_IDENT

TOKEN_INTEGER

TOKEN_INTERPOL_END

TOKEN_INTERPOL_START

TOKEN_PATH

TOKEN_URI

TOKEN_STRING_CONTENT

TOKEN_STRING_END

TOKEN_STRING_START

NODE_APPLY

NODE_ASSERT

NODE_ATTRPATH

NODE_DYNAMIC

NODE_ERROR

NODE_IDENT

NODE_IF_ELSE

NODE_SELECT

NODE_INHERIT

NODE_INHERIT_FROM

NODE_STRING

NODE_INTERPOL

NODE_LAMBDA

NODE_IDENT_PARAM

NODE_LEGACY_LET

NODE_LET_IN

NODE_LIST

NODE_BIN_OP

NODE_PAREN

NODE_PATTERN

NODE_PAT_BIND

NODE_PAT_ENTRY

NODE_ROOT

NODE_ATTR_SET

NODE_ATTRPATH_VALUE

NODE_UNARY_OP

NODE_LITERAL

NODE_WITH

NODE_PATH

NODE_HAS_ATTR

Implementations

impl SyntaxKind

pub fn is_literal(self) -> bool

Returns true if this token is a literal, such as an integer or a string

pub fn is_fn_arg(self) -> bool

Returns true if this token should be used as a function argument.

Example:
add 1 2 + 3
^   ^ ^ ^
|   | | +- false
|   | +- true
|   +- true
+- true

pub fn is_trivia(self) -> bool

Returns true if this token is a comment, whitespace, or similar, and should be skipped over by the parser.